Move tests for deleting in progress submission to test_profile.py
